Move HexString from Wiring to SmingCore/Data
