+++ Sming/SmingCore/Data/HexString.cpp
