+++ Sming/SmingCore/Data/HexString.h
