+++ Sming/SmingCore/Network/SmtpClient.cpp
@@ -20,7 +20,7 @@
 #include "../Services/WebHelpers/base64.h"
 #include "Data/Stream/QuotedPrintableOutputStream.h"
 #include "Data/Stream/Base64OutputStream.h"
-#include "HexString.h"
+#include "Data/HexString.h"
 
 #if !defined(ENABLE_SSL) || ENABLE_SSL == 0
 // if our SSL is not used then we try to use the one coming from the SDK
