+++ Sming/SmingCore/Platform/Station.cpp
@@ -7,7 +7,7 @@
 
 #include "Station.h"
 #include "../../SmingCore/SmingCore.h"
-#include "HexString.h"
+#include "Data/HexString.h"
 
 StationClass WifiStation;
